+++ app/Http/Controllers/Backend/SubCategoryController.php
@@ -8,6 +8,7 @@ use App\DataTables\SubCategoryDataTable;
 use App\Http\Controllers\Controller;
 use App\Models\SubCategory;
 use App\Models\Category;
+use App\Models\ChildCategory;
 use Illuminate\Http\Request;
 use Str;
 
@@ -96,6 +97,11 @@ class SubCategoryController extends Controller
     public function destroy(string $id)
     {
         $subCategory=SubCategory::findOrFail($id);
+        $childCategory= ChildCategory::where('sub_category_id',$subCategory->id)->count();
+
+        if($childCategory>0){
+            return response(['status'=> 'error','message'=>'This items contain sub items for delete this you have to delete the sub items first']);
+        }
         $subCategory->delete();
 
         return response(['status'=>'success','message'=>'Deleted Successful!']);
